Added beam momentum
Flashlight beam now trails mouse movement a little (beam is targeted not at [mouseX, mouseY] but at an invisible object which follows the mouse around at a fixed speed). This both gives it a bit more heft, and also makes it much easier for us to 'steal' the flashlight beam away from the player when needed.